Housekeeping and moving gnderchart to module for future reference
