+++ src/index.js
@@ -573,44 +573,6 @@ function drawKpis(totals, totalsDiff) {
   );
 }
 
-var genderChart = false;
-function drawGenderChart(gender) {
-  var options = {
-    series: [gender.female, gender.male, gender.unspecified],
-    colors: ["#57a998", "#c8e6bc", "#85cb9c"],
-    chart: {
-      width: 400,
-      type: "pie",
-    },
-    labels: [
-      i18next.t("Female"),
-      i18next.t("Male"),
-      i18next.t("UnspecifiedGender"),
-    ],
-    responsive: [
-      {
-        breakpoint: 480,
-        options: {
-          chart: {
-            width: 400,
-          },
-          legend: {
-            position: "bottom",
-          },
-        },
-      },
-    ],
-  };
-  if (genderChart !== false) {
-    genderChart.destroy();
-  }
-  genderChart = new ApexCharts(
-    document.querySelector("#gender-chart"),
-    options
-  );
-  genderChart.render();
-}
-
 /**
  * @param {string} lastUpdated - MMM DD YYYY, HH:mm JST (e.g. Mar 29 2020, 15:53 JST)
  */
@@ -864,7 +826,7 @@ function setLang(lng) {
       hotspotMap = drawHotspotMap(ddb.prefectures, LANG);
 
       drawPrefectureTrajectoryChart(ddb.prefectures);
-      //drawGenderChart(ddb.gender);
+
       drawAgeTrendChart(ddb.age);
 
       dailyIncreaseChart = drawDailyIncreaseChart(
@@ -951,7 +913,6 @@ function loadDataOnPage() {
       drawDailyIncreaseChart(ddb.trend);
       drawPrefectureTrajectoryChart(ddb.prefectures);
       drawAgeTrendChart(ddb.age);
-      // drawGenderChart(ddb.gender);
       testingTrendChart = drawTestingTrendChart(ddb.trend, testingTrendChart);
       dailyIncreaseChart = drawDailyIncreaseChart(
         ddb.trend,
